requirements.txt
azure-identity==1.14.0
azure-storage-blob==12.19.0
openai==1.12.0
pydantic==2.5.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.23.2
# asyncio is in the stdlib; avoid PyPI package
aiohttp==3.9.5
requests==2.31.0
azure-functions==1.18.0
applicationinsights==0.11.10
